requirements.txt
python==3.10.9
numpy==1.24.2
Pillow==10.3.0
scikit_learn==1.1.1
scipy==1.10.0
torch==1.12.0
torchvision==0.13.0
tqdm==4.64.0
